The initial branch for init.ps1 is now "main"
Fixes #1

+++ init.ps1
@@ -1,5 +1,11 @@
 #!/usr/bin/env pwsh
 
+$gitVersion = & git version
+if ($gitVersion -lt 'git version 2.28.0') {
+    Write-Host -ForegroundColor Red 'You need at least git version 2.28'
+    exit 1
+}
+
 Write-Host -ForegroundColor Cyan 'Removing ".git"...'
 Remove-Item "$PSScriptRoot/.git" -Recurse -Force
 
@@ -19,7 +25,7 @@ Remove-Item "$PSScriptRoot/docs" -Recurse -Force
 
 Write-Host
 Write-Host -ForegroundColor Cyan 'Initializing new Git repository...'
-& git init "$PSScriptRoot"
+& git init --initial-branch=main "$PSScriptRoot"
 
 Write-Host
 Write-Host -ForegroundColor Cyan 'Adding skeleton files...'
